Fix error stuff. Remove deprecated ui elements. Add time_size_label

+++ hfbm/Triangulation/Delaunay.cpp
@@ -3,7 +3,7 @@
 #include <stdexcept>
 #include <iostream>
 
-Delaunay::Delaunay(std::unique_ptr<HeightMap> heightMap, int meshHeight, double error) : 
+Delaunay::Delaunay(std::unique_ptr<HeightMap> heightMap, double meshHeight, double error) : 
 	Greedy(std::move(heightMap), meshHeight, error)
 {}
